+++ lib/src/ui/infinite_scroll_view.dart
@@ -76,9 +76,9 @@ class _RenderInfiniteScrollView extends RenderShiftedBox {
   ViewportOffset _position;
   set position(ViewportOffset value) {
     if (_position == value) return;
-    if (attached) _position.removeListener(markNeedsLayout);
+    if (attached) _position.removeListener(_onScroll);
     _position = value;
-    if (attached) _position.addListener(markNeedsLayout);
+    if (attached) _position.addListener(_onScroll);
     markNeedsLayout();
   }